Use threads to pass data back and forth
This method is not going to work because
Urth code cannot give input to threads.
So this is just a checkpoint before switching
methods, so that, if in the future threads
can accept input from Urth, I'm not starting
from scratch when I re-implement.

In this commit, the client (either Urth or
Mars frontent) pokes the client Gall app,
which spawns one thread to handle sending
audio data to the provider ship and
a second thread to receive responses from it.

The provider Gall app is poked, itself
spawning two threads: one to receive audio
from the client thread and the second to send
JSON transcripts to the other client thread,
from the Urth backend and transcription Engine.

Since this won't work because Urth cannot
give input to threads, the plan is to instead
pipe input from the client frontend through
the client Gall app (and similarly for the
provider backend through the provider Gall app).

The client Gall app will then send the audio
to a provider thread which will be subscribed
to it; similarly, the provider Gall app will
send the JSON transcripts to a client thread
subscribed to it.