Use Gall apps to send
Current architecture is:

* Relay audio through client Gall app via pokes,
  to provider thread via subscription,
  and finally to Golang Urth backend via subscription.
* Relay transcripts through provider Gall app
  via pokes, to client thread via subscription,
  and finally to frontend via subscription.

Code currently succeeds in passing options and
some corrupted form of audio bytes from frontend,
through Mars, to the Golang Urth backend, and to
the Engine.
The Golang Urth backend receives `[noise]`
transcripts from the Engine, so the audio is
getting messed up somewhere along the line.
The transcripts are not currently properly
propagated back through Mars.

In addition, things are very slow.
My hope is that this is because I am spamming
tons of pokes to send the audio, and that
optimizing the size of the audio chunks can
reduce the lag.
My fear is that the Urbit protocols are simply not
currently designed/capable of passing low latency
streaming data.

Worst comes to worst, once I fix the audio bugs,
it'll just be laggy (unless pokes are getting
processed out of order -- which would be very bad).

Aside from fixing bugs and properly propagating
the transcripts, other TODOs are:

1. Stop sending audio gracefully.
2. Shutdown threads, etc., gracefully when done.
3. JS frontend: poke some graph to store
   transcripts in a `%chat` graph.
4. Performance tuning.
5. Error handling.